test: Split tests between `create` and `extract_mount`
