+++ arx/tests/create.rs
@@ -115,125 +115,3 @@ fn test_crate_existant_output() {
         assert_eq!(&buf, b"jbkC\x00\x00\x00\x00\x00\x02");
     }
 }
-
-#[cfg(all(unix, not(feature = "in_ci")))]
-#[test]
-fn test_create_and_mount() {
-    let tmp_source_dir = SHARED_TEST_DIR.path();
-    let arx_file = BASE_ARX_FILE.path();
-
-    let mount_point = tempfile::TempDir::new_in(env!("CARGO_TARGET_TMPDIR")).unwrap();
-    let arx = arx::Arx::new(arx_file).unwrap();
-    let arxfs = arx::ArxFs::new(arx).unwrap();
-    let _mount_handle = arxfs
-        .spawn_mount("Test mounted arx".into(), mount_point.path())
-        .unwrap();
-    let output = cmd!("diff", "-r", tmp_source_dir, mount_point.path());
-    println!("Out : {}", String::from_utf8(output.stdout).unwrap());
-    println!("Err: {}", String::from_utf8(output.stderr).unwrap());
-    assert!(output.status.success());
-}
-
-#[test]
-fn test_create_and_extract() {
-    let tmp_source_dir = SHARED_TEST_DIR.path();
-    let arx_file = BASE_ARX_FILE.path();
-
-    let extract_dir = tempfile::TempDir::new_in(env!("CARGO_TARGET_TMPDIR")).unwrap();
-    arx::extract(
-        &arx_file,
-        extract_dir.path(),
-        Default::default(),
-        true,
-        false,
-    )
-    .unwrap();
-    let output = cmd!("diff", "-r", tmp_source_dir, extract_dir.path());
-    println!("Out : {}", String::from_utf8(output.stdout).unwrap());
-    println!("Err: {}", String::from_utf8(output.stderr).unwrap());
-    assert!(output.status.success());
-}
-
-#[test]
-fn test_create_and_extract_filter() {
-    let tmp_source_dir = SHARED_TEST_DIR.path();
-    let arx_file = BASE_ARX_FILE.path();
-
-    let extract_dir =
-        tempfile::TempDir::with_prefix_in("extract_", env!("CARGO_TARGET_TMPDIR")).unwrap();
-    arx::extract(
-        &arx_file,
-        extract_dir.path(),
-        ["sub_dir_a".into()].into(),
-        true,
-        true,
-    )
-    .unwrap();
-
-    let mut source_sub_dir = tmp_source_dir.to_path_buf();
-    source_sub_dir.push("sub_dir_a");
-    let mut extract_sub_dir = extract_dir.path().to_path_buf();
-    extract_sub_dir.push("sub_dir_a");
-
-    println!(
-        "Diff {} and {}",
-        source_sub_dir.display(),
-        extract_sub_dir.display()
-    );
-    let output = cmd!("diff", "-r", &source_sub_dir, &extract_sub_dir);
-    println!("Out : {}", String::from_utf8(output.stdout).unwrap());
-    println!("Err: {}", String::from_utf8(output.stderr).unwrap());
-    assert!(output.status.success());
-}
-
-#[test]
-fn test_create_and_extract_subdir() {
-    let tmp_source_dir = SHARED_TEST_DIR.path();
-    let arx_file = BASE_ARX_FILE.path();
-
-    let extract_dir =
-        tempfile::TempDir::with_prefix_in("extract_", env!("CARGO_TARGET_TMPDIR")).unwrap();
-
-    let output = cmd!(
-        "arx",
-        "extract",
-        &arx_file,
-        "--root-dir",
-        "sub_dir_a",
-        "-C",
-        extract_dir.path()
-    );
-    assert!(output.status.success());
-
-    let mut source_sub_dir = tmp_source_dir.to_path_buf();
-    source_sub_dir.push("sub_dir_a");
-
-    println!(
-        "Diff {} and {}",
-        source_sub_dir.display(),
-        extract_dir.path().display()
-    );
-    let output = cmd!("diff", "-r", &source_sub_dir, extract_dir.path());
-    println!("Out: {}", String::from_utf8(output.stdout).unwrap());
-    println!("Err: {}", String::from_utf8(output.stderr).unwrap());
-    assert!(output.status.success());
-}
-
-#[test]
-fn test_create_and_extract_subfile() {
-    let arx_file = BASE_ARX_FILE.path();
-
-    let extract_dir =
-        tempfile::TempDir::with_prefix_in("extract_", env!("CARGO_TARGET_TMPDIR")).unwrap();
-
-    let output = cmd!(
-        "arx",
-        "extract",
-        &arx_file,
-        "--root-dir",
-        "sub_dir_a/file1.txt",
-        "-C",
-        extract_dir.path()
-    );
-    assert!(!output.status.success());
-}
